+++ vnext/Desktop/pch.h
@@ -13,4 +13,7 @@
 #include <winrt/base.h>
 #include <mutex>
 
+#include <winrt/Windows.Foundation.Collections.h>
+#include <winrt/Windows.Foundation.h>
+
 #include <Base/CxxReactIncludes.h>