+++ vnext/Microsoft.ReactNative.IntegrationTests/TurboModuleTests.cpp
@@ -184,7 +184,6 @@ TEST_CLASS (TurboModuleTests) {
 
     auto reactNativeHost = TestReactNativeHostHolder(L"TurboModuleTests", [](ReactNativeHost const &host) noexcept {
       host.PackageProviders().Append(winrt::make<SampleTurboModulePackageProvider>());
-      return true;
     });
 
     TestEventService::ObserveEvents({