+++ vnext/Microsoft.ReactNative/Utils/XamlIslandUtils.cpp
@@ -47,13 +47,18 @@ void FixProofingMenuCrashForXamlIsland(xaml::Controls::Primitives::FlyoutBase co
               }
             });
           } else if (appBarButton.Flyout() == textBox.ProofingMenuFlyout()) {
-            // Replace the AppBarButton for the proofing menu with one that doesn't crash
-            const auto customAppBarButton = winrt::make<CustomAppBarButton>();
-            customAppBarButton.Label(appBarButton.Label());
-            customAppBarButton.Icon(appBarButton.Icon());
-            customAppBarButton.Flyout(appBarButton.Flyout());
-            commands.RemoveAt(i);
-            commands.InsertAt(i, customAppBarButton);
+            if (!appBarButton.try_as<CustomAppBarButton>()) {
+              // Replace the AppBarButton for the proofing menu with one that doesn't crash
+              const auto customAppBarButton = winrt::make<CustomAppBarButton>();
+              customAppBarButton.Label(appBarButton.Label());
+              customAppBarButton.Icon(appBarButton.Icon());
+              customAppBarButton.Flyout(appBarButton.Flyout());
+              commands.RemoveAt(i);
+              commands.InsertAt(i, customAppBarButton);
+            } else if (!textBox.IsSpellCheckEnabled()) {
+              // Remove proofing menu option if spell-check is disabled
+              commands.RemoveAt(i);
+            }
 
             // There is only one proofing menu option
             break;