+++ vnext/src/Libraries/Components/Button.windows.js
@@ -355,7 +355,11 @@ class Button extends React.Component<
           onPress={onPress}
           tabIndex={tabIndex}
           touchSoundDisabled={touchSoundDisabled}
-          underlayColor={PlatformColor('ButtonBackgroundPressed')}
+          underlayColor={
+            color
+              ? PlatformColor('SolidBackgroundFillColorBaseBrush')
+              : PlatformColor('ButtonBackgroundPressed')
+          }
           onShowUnderlay={() => {
             this.setState({pressed: true});
           }}
@@ -363,24 +367,12 @@ class Button extends React.Component<
             this.setState({pressed: false});
           }}
           style={
-            this.state.pressed
-              ? [
-                  buttonStyles,
-                  {
-                    borderColor: PlatformColor('ButtonBorderBrushPressed'),
-                    borderBottomWidth: 1,
-                  },
-                ]
+            color
+              ? {borderRadius: 3}
+              : this.state.pressed
+              ? [buttonStyles, styles.buttonPressed]
               : this.state.hover
-              ? [
-                  buttonStyles,
-                  {
-                    backgroundColor: PlatformColor(
-                      'ButtonBackgroundPointerOver',
-                    ),
-                    borderColor: PlatformColor('ButtonBorderBrushPointerOver'),
-                  },
-                ]
+              ? [buttonStyles, styles.buttonHover]
               : buttonStyles
           }
           onMouseEnter={() => {
@@ -389,27 +381,52 @@ class Button extends React.Component<
           onMouseLeave={() => {
             if (!disabled) this.setState({hover: false});
           }}>
-          <Text
+          <View
             style={
-              this.state.pressed
-                ? [
-                    textStyles,
-                    {
-                      color: PlatformColor('ButtonForegroundPressed'),
-                    },
-                  ]
-                : this.state.hover
-                ? [
-                    textStyles,
-                    {
-                      color: PlatformColor('ButtonForegroundPointerOver'),
-                    },
-                  ]
-                : textStyles
-            }
-            disabled={disabled}>
-            {formattedTitle}
-          </Text>
+              color
+                ? this.state.pressed
+                  ? [buttonStyles, styles.buttonPressed]
+                  : buttonStyles
+                : {}
+            }>
+            <View
+              style={
+                color
+                  ? this.state.hover
+                    ? [
+                        styles.buttonHover,
+                        {
+                          backgroundColor: 'rgba(255, 255, 255, 0.1)',
+                        },
+                      ]
+                    : {}
+                  : {}
+              }>
+              <Text
+                style={
+                  color
+                    ? textStyles
+                    : this.state.pressed
+                    ? [
+                        textStyles,
+                        {
+                          color: PlatformColor('ButtonForegroundPressed'),
+                        },
+                      ]
+                    : this.state.hover
+                    ? [
+                        textStyles,
+                        {
+                          color: PlatformColor('ButtonForegroundPointerOver'),
+                        },
+                      ]
+                    : textStyles
+                }
+                disabled={disabled}>
+                {formattedTitle}
+              </Text>
+            </View>
+          </View>
         </Touchable>
       );
     } else {
@@ -509,6 +526,17 @@ const styles = StyleSheet.create({
     },
     // Windows]
   }),
+  // [Windows
+  buttonHover: {
+    backgroundColor: PlatformColor('ButtonBackgroundPointerOver'),
+    borderColor: PlatformColor('ButtonBorderBrushPointerOver'),
+    borderRadius: 3,
+  },
+  buttonPressed: {
+    borderColor: PlatformColor('ButtonBorderBrushPressed'),
+    borderBottomWidth: 1,
+  },
+  // Windows]
 });
 
 module.exports = Button;